Move extension runtime to separate crate
