+++ core/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod command;
 pub mod config;
 pub mod crypto;
-pub mod extension;
 pub mod instance;
 pub mod level;
 pub mod network;
