+++ ext/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "pyro-extensions"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.70"
+async-trait = "0.1.68"
+flate2 = "1.0.25"
+is-terminal = "0.4.6"
+sha2 = "0.10.6"
+tracing = "0.1.37"
+wasi-common = "7.0.0"
+wasmtime = "7.0.0"
+wasmtime-wasi = "7.0.0"
