+++ ext/src/cache.rs
