+++ level/test/db/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-000802
+MANIFEST-000804
